feat: Rename internal member cluster package. Minor styling refactor. (#52)
