+++ pkg/controllers/internalMemberCluster/member_controller_test.go
@@ -1,4 +1,4 @@
-package memberinternalmembercluster
+package internalMemberCluster
 
 import (
 	"testing"
